+++ front/src/components/atoms/index.ts
@@ -1,13 +1,11 @@
 import { Select } from './Select'
 import { Spinner } from './Spinner'
 import { Title } from './Title'
-import Button from './Button'
 import Dropdown from './Dropdown'
 import CheckBox from './CheckBox'
 import Icon from './Icon'
 import Input from './Input'
 import Label from './Label'
-import Link from './Link'
 import Radio from './Radio'
 import Text from './Text'
 import Textarea from './Textarea'
@@ -17,13 +15,11 @@ import UserImage from './UserImage'
 import Counter from './Counter'
 
 export {
-  Button,
   Dropdown,
   CheckBox,
   Icon,
   Input,
   Label,
-  Link,
   Radio,
   Select,
   Spinner,
@@ -33,5 +29,5 @@ export {
   ValidationMessage,
   HamburgerMenu,
   UserImage,
-  Counter
+  Counter,
 }