+++ packages/ui/rollup.config.js
@@ -4,32 +4,41 @@ import typescript from '@rollup/plugin-typescript'
 import terser from '@rollup/plugin-terser'
 import external from 'rollup-plugin-peer-deps-external'
 import postcss from 'rollup-plugin-postcss'
+import { dts } from 'rollup-plugin-dts'
 import image from '@rollup/plugin-image'
 
 const pkg = require('./package.json')
 
-export default {
-  input: 'src/index.ts',
-  output: [
-    {
-      file: pkg.main,
-      format: 'cjs',
-      sourcemap: true,
-      name: 'react-lib',
-    },
-    {
-      file: pkg.module,
-      format: 'esm',
-      sourcemap: true,
-    },
-  ],
-  plugins: [
-    external(),
-    image(),
-    resolve(),
-    commonjs(),
-    typescript({ tsconfig: './tsconfig.build.json' }),
-    postcss(),
-    terser(),
-  ],
-}
+export default [
+  {
+    input: 'src/index.ts',
+    output: [
+      {
+        file: pkg.main,
+        format: 'cjs',
+        sourcemap: true,
+      },
+      {
+        file: pkg.module,
+        format: 'esm',
+        sourcemap: true,
+      },
+    ],
+    plugins: [
+      external(),
+      image(),
+      resolve(),
+      commonjs(),
+      typescript({ tsconfig: './tsconfig.build.json' }),
+      postcss(),
+      terser(),
+    ],
+    external: ['react', 'react-dom', 'styled-components'],
+  },
+  {
+    input: 'dist/esm/index.d.ts',
+    output: [{ file: 'dist/index.d.ts', format: 'esm' }],
+    plugins: [dts()],
+    external: ['react', 'react-dom', 'styled-components'],
+  },
+]